+++ activerecord/lib/active_record/connection_adapters/abstract/schema_definitions.rb
@@ -887,6 +887,15 @@ module ActiveRecord
         @base.remove_check_constraint(name, *args, **options)
       end
 
+      # Checks to see if a check constraint exists on a table.
+      #
+      #  t.remove_check_constraint(name: "price_check") unless t.check_constraint_exists?(name: "price_check")
+      #
+      # See {connection.check_constraint_exists?}[rdoc-ref:SchemaStatements#check_constraint_exists?]
+      def check_constraint_exists?(*args, **options)
+        @base.check_constraint_exists?(name, *args, **options)
+      end
+
       private
         def raise_on_if_exist_options(options)
           unrecognized_option = options.keys.find do |key|