+++ activerecord/lib/active_record/connection_adapters/abstract/schema_statements.rb
@@ -1231,11 +1231,18 @@ module ActiveRecord
       #
       #   remove_check_constraint :products, name: "price_check"
       #
+      # Checks if the check constraint exists before trying to remove it. Will silently ignore check constraint that
+      # don't exist.
+      #
+      #   remove_check_constraint :products, name: "price_check", if_exists: true
+      #
       # The +expression+ parameter will be ignored if present. It can be helpful
       # to provide this in a migration's +change+ method so it can be reverted.
       # In that case, +expression+ will be used by #add_check_constraint.
       def remove_check_constraint(table_name, expression = nil, **options)
         return unless supports_check_constraints?
+
+        return if options[:if_exists] && !check_constraint_exists?(table_name, **options)
 
         chk_name_to_delete = check_constraint_for!(table_name, expression: expression, **options).name
 
@@ -1243,6 +1250,14 @@ module ActiveRecord
         at.drop_check_constraint(chk_name_to_delete)
 
         execute schema_creation.accept(at)
+      end
+
+      # Checks to see if a check constraint exists on a table.
+      #
+      #   # Checks if a check constraint exists
+      #   check_constraint_exists?(:products, name: "price_check")
+      def check_constraint_exists?(table_name, **options)
+        check_constraint_for(table_name, **options).present?
       end
 
       def dump_schema_information # :nodoc: