+++ activerecord/CHANGELOG.md
@@ -1,3 +1,13 @@
+*   Introduce `check_constraint_exists?` method which checks
+    if a check constraint exists on a table.
+
+    Add support for `if_exists` option for removing a check constraint.
+
+    The `remove_check_constraint` method can take an `if_exists` option. If this is set
+    to true an error won't be raised if the check constraint doesn't exist.
+
+    *Aditya Bhutani and Margaret Parsa*
+
 *   `find_or_create_by` now try to find a second time if it hits a unicity constraint.
 
     `find_or_create_by` always has been inherently racy, either creating multiple